python_backend/requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.12
langchain-huggingface==0.0.1
chromadb==0.4.22
firecrawl-py==0.0.8
fastapi==0.108.0
uvicorn==0.25.0
python-dotenv==1.0.0
pydantic==2.5.0
sentence-transformers==2.2.2
numpy==1.24.3
python-multipart==0.0.6
</parameter>